refactor(analyzer): Move option key name constants to companion objects
Apply proper scoping to avoid potential constant name clashes.
