+++ analyzer/src/main/kotlin/managers/Gradle.kt
@@ -61,11 +61,6 @@ import org.ossreviewtoolkit.utils.common.splitOnWhitespace
 import org.ossreviewtoolkit.utils.common.temporaryProperties
 import org.ossreviewtoolkit.utils.ort.createOrtTempFile
 
-/**
- * The name of the option to specify the Gradle version.
- */
-const val OPTION_GRADLE_VERSION = "gradleVersion"
-
 private val GRADLE_USER_HOME = Os.env["GRADLE_USER_HOME"]?.let { File(it) } ?: Os.userHomeDirectory.resolve(".gradle")
 
 private val GRADLE_BUILD_FILES = listOf("build.gradle", "build.gradle.kts")
@@ -87,7 +82,12 @@ class Gradle(
     analyzerConfig: AnalyzerConfiguration,
     repoConfig: RepositoryConfiguration
 ) : PackageManager(name, analysisRoot, analyzerConfig, repoConfig) {
-    companion object : Logging
+    companion object : Logging {
+        /**
+         * The name of the option to specify the Gradle version.
+         */
+        const val OPTION_GRADLE_VERSION = "gradleVersion"
+    }
 
     class Factory : AbstractPackageManagerFactory<Gradle>("Gradle") {
         // Gradle prefers Groovy ".gradle" files over Kotlin ".gradle.kts" files, but "build" files have to come before
