+++ analyzer/src/main/kotlin/managers/Npm.kt
@@ -83,9 +83,6 @@ import org.ossreviewtoolkit.utils.common.withoutPrefix
 import org.semver4j.RangesList
 import org.semver4j.RangesListFactory
 
-/** Name of the configuration option to toggle legacy peer dependency support. */
-const val OPTION_LEGACY_PEER_DEPS = "legacyPeerDeps"
-
 /** Name of the scope with the regular dependencies. */
 private const val DEPENDENCIES_SCOPE = "dependencies"
 
@@ -110,7 +107,10 @@ open class Npm(
     analyzerConfig: AnalyzerConfiguration,
     repoConfig: RepositoryConfiguration
 ) : PackageManager(name, analysisRoot, analyzerConfig, repoConfig), CommandLineTool {
-    companion object : Logging
+    companion object : Logging {
+        /** Name of the configuration option to toggle legacy peer dependency support. */
+        const val OPTION_LEGACY_PEER_DEPS = "legacyPeerDeps"
+    }
 
     class Factory : AbstractPackageManagerFactory<Npm>("NPM") {
         override val globsForDefinitionFiles = listOf("package.json")
