+++ analyzer/src/main/kotlin/managers/Pip.kt
@@ -44,11 +44,9 @@ object Python : CommandLineTool {
     override fun transformVersion(output: String) = output.removePrefix("Python ")
 }
 
-const val OPTION_OPERATING_SYSTEM = "operatingSystem"
 private const val OPTION_OPERATING_SYSTEM_DEFAULT = "linux"
 private val OPERATING_SYSTEMS = listOf(OPTION_OPERATING_SYSTEM_DEFAULT, "macos", "windows")
 
-const val OPTION_PYTHON_VERSION = "pythonVersion"
 private const val OPTION_PYTHON_VERSION_DEFAULT = "3.10"
 private val PYTHON_VERSIONS = listOf("2.7", "3.6", "3.7", "3.8", "3.9", OPTION_PYTHON_VERSION_DEFAULT)
 
@@ -68,7 +66,10 @@ class Pip(
     analyzerConfig: AnalyzerConfiguration,
     repoConfig: RepositoryConfiguration
 ) : PackageManager(name, analysisRoot, analyzerConfig, repoConfig) {
-    companion object : Logging
+    companion object : Logging {
+        const val OPTION_OPERATING_SYSTEM = "operatingSystem"
+        const val OPTION_PYTHON_VERSION = "pythonVersion"
+    }
 
     class Factory : AbstractPackageManagerFactory<Pip>("PIP") {
         override val globsForDefinitionFiles = listOf("*requirements*.txt", "setup.py")
